+++ docs/whats-new.rst
@@ -15,7 +15,14 @@ Coming up on the next release
 
 - **New 'research' user mode**. This new feature implements automatic filtering of Argo data following international recommendations for research/climate studies. With this user mode, only Delayed Mode with good QC data are returned. Check out the :ref:`user-mode` section for all the details.
 
-- **Argopy now provides authenticated access to the Argo reference database for DMQC**. Using user/password new **argopy** options, it is now possible to fetch the `Argo CTD reference database <http://www.argodatamgt.org/DMQC/Reference-data-base/Latest-Argo-Reference-DB>`_, with the :class:`CTDRefDataFetcher` class. (:pr:`256`) by `G. Maze <http://www.github.com/gmaze>`_
+- **argopy now provides a specific xarray engine to properly read Argo netcdf files**. Using ``engine='argo'`` in :func:`xarray.open_dataset`, all variables will properly be casted, i.e. returned with their expected data types, which is not the case otherwise. This works with *ALL* Argo netcdf file types (as listed in the `Reference table R01 <http://vocab.nerc.ac.uk/collection/R01/current/>`_).  (:pr:`208`) by `G. Maze <http://www.github.com/gmaze>`_
+
+.. code-block:: python
+
+    import xarray as xr
+    ds = xr.open_dataset("dac/aoml/1901393/1901393_prof.nc", engine='argo')
+
+- **argopy now provides authenticated access to the Argo reference database for DMQC**. Using user/password new **argopy** options, it is now possible to fetch the `Argo CTD reference database <http://www.argodatamgt.org/DMQC/Reference-data-base/Latest-Argo-Reference-DB>`_, with the :class:`CTDRefDataFetcher` class. (:pr:`256`) by `G. Maze <http://www.github.com/gmaze>`_
 
 .. code-block:: python
 
@@ -24,6 +31,10 @@ Coming up on the next release
     with argopy.set_options(user="john_doe", password="***"):
         f = CTDRefDataFetcher(box=[15, 30, -70, -60, 0, 5000.0])
         ds = f.to_xarray()
+
+- New option to control the expiration time of cache file ``cache_expiration``. 
+
+**Internals**
 
 - **Index store can now export search results to standard Argo index file format**. See all details in :ref:`Store: Low-level Argo Index access`. (:pr:`260`) by `G. Maze <http://www.github.com/gmaze>`_
 
@@ -48,13 +59,9 @@ Coming up on the next release
     idx = indexstore(index_file="argo_bio-profile_index.txt").load()
     idx.search_params(['C1PHASE_DOXY', 'DOWNWELLING_PAR'])
 
-- New option to control the expiration time of cache file ``cache_expiration``. 
-
-**Internals**
-
 - Use a mocked server for all http and GDAC ftp requests in CI tests (:pr:`249`, :pr:`252`, :pr:`255`) by `G. Maze <http://www.github.com/gmaze>`_
 - Removed support for minimal dependency requirements and for python 3.7. (:pr:`252`) by `G. Maze <http://www.github.com/gmaze>`_
-- Changed License from Apache to `CeCILL 2.1 <http://www.cecill.info/licences/Licence_CeCILL_V2.1-en.html>`_ (:commit:`3eadce73bab1c4bd1a770af196fc8bd1126e8ad2`)
+- Changed License from Apache to `EUPL 1.2 <https://opensource.org/license/eupl-1-2>`_
 
 v0.1.13 (28 Mar. 2023)
 ----------------------
@@ -62,6 +69,7 @@ v0.1.13 (28 Mar. 2023)
 **Features and front-end API**
 
 - **New utility class to retrieve the Argo deployment plan from the Ocean-OPS api.** This is the utility class :class:`OceanOPSDeployments`. See the new documentation section on :ref:`Deployment Plan` for more. (:pr:`244`) by `G. Maze <http://www.github.com/gmaze>`_
+
 
 .. code-block:: python
 
@@ -116,6 +124,10 @@ v0.1.13 (28 Mar. 2023)
 **Breaking changes**
 
 - Data source ``localftp`` is deprecated and removed from **argopy**. It's been replaced by the ``gdac`` data source with the appropriate ``ftp`` option. See :ref:`Data sources`. (:pr:`240`) by `G. Maze <http://www.github.com/gmaze>`_
+
+**Breaking changes with previous versions**
+
+- :class:`argopy.utilities.ArgoNVSReferenceTables` methods ``all_tbl`` and ``all_tbl_name`` are now properties, not methods.
 
 
 v0.1.12 (16 May 2022)